done GetMonthSummary to get different month's summary
